feat(devtools): rename LocalModePage component
